backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Mikołaj Garbowski", email = "mikolaj.garbowski@gmail.com" },
]
dependencies = ["fastapi[all]>=0.116.1", "asyncpg>=0.30.0", "sqlalchemy[asyncio]>=2.0.43", "python-multipart>=0.0.20", "numpy>=2.3.2", "pgvector>=0.4.1", "facenet-pytorch>=2.5.3", "torch>=2.8.0", "pillow>=11.3.0", "pydantic-settings>=2.10.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = { shell = "fastapi dev", help = "Run the development server with auto-reload" }
